Opened Repository with basic project structure.
I began working on this project without first creating a repository. Therefore, this initial upload includes the basic structure
of the frontend and backend, including a test frontend page that calls on the backend and a skeletal backend framework. This
commit includes all the package files for both react and express/node. In the world-web-backend folder, it includes a routes folder (with routes),
a controllers folder (with middleware), and a db folder (for queries and connections) as well as an app.js function that acts as the main file.
The front end world-web folder (using the React framework) includes within source a components folder with components, a routes folder
with routes, and a styles folder with styles. A lot of this structure is subject to change as this project inevtiably expands and I
learn more about this process.