Added navbar
This commit adds a navbar that the user can use to navigate the page (still WIP). This file
was previously named 'header', but was renamed to navbar. It also added an accomanying
navbar.modules.css style element. Alterations were made to app.jsx to display the navbar on the homepage.