Added basic forum functionality.
The forum link is no longer broken, and will now load a page that allows users to see all
posts and if they are logged in, create their own post and reply to other posts.
Currently users cannot edit or delete their own posts and post authors do not load
correctly in the detail view. But the basics are there.